created audio-bar component
